Replace old Hero file
