+++ app/pages/hero.tsx
@@ -1,69 +1,195 @@
 "use client";
-
+import { useRef, useLayoutEffect, useState } from "react";
 import Image from "next/image";
-import { useState, useEffect, useRef } from "react";
-import StarryBackground from "../components/StarryBackground";
-import Navbar from "../components/navbar";
-import { initHeroAnimations } from "../animations/heroAnimations";
+import Stars from "../components/stars";
+import { gsap } from "gsap";
 
 export default function Hero() {
-  const [skylineVisible, setSkylineVisible] = useState(false);
-  const heroRef = useRef<HTMLElement>(null);
-  const logoRef = useRef<HTMLDivElement>(null);
+  const comp = useRef<HTMLElement>(null);
+  const logoRef = useRef<HTMLImageElement>(null);
   const titleRef = useRef<HTMLDivElement>(null);
+  const preTitleRef = useRef<HTMLParagraphElement>(null);
+  const subTitleRef = useRef<HTMLParagraphElement>(null);
   const skylineRef = useRef<HTMLDivElement>(null);
+  const [clickCount, setClickCount] = useState(0);
 
-  useEffect(() => {
-    initHeroAnimations(logoRef, titleRef, skylineRef, setSkylineVisible);
+  useLayoutEffect(() => {
+    let ctx = gsap.context(() => {
+      const tl = gsap.timeline();
+      const isMobile = window.innerWidth < 768;
+
+      if (!isMobile) {
+        // Spin Logo with fade
+        tl.to(logoRef.current, { scale: 1, rotation: 360, opacity: 1, duration: 1.2, ease: "back.out(1.7)", delay: 1 });
+
+        // Slide from left, wipe
+        tl.fromTo(titleRef.current,
+          { x: -600, clipPath: "inset(0% 0% 0% 100%)", opacity: 1},
+          { x: 0, clipPath: "inset(0% 0% 0% 0%)", duration: 1, ease: "power3.out" },
+        "-=1.2");
+
+      } else {
+        // Comet flies in
+        tl.fromTo(logoRef.current,
+          { x: 100, y: -100, opacity: 0, scale: 0.8 },
+          { x:0, y: 0, opacity: 1, scale: 1, duration: 1.2, ease: "power3.out", delay: 1 },
+        );
+        tl.fromTo(titleRef.current,
+          { y: 50, opacity: 0 },
+          { y: 0, opacity: 1, duration: 1, ease: "power3.out" },
+        "-=1.0");
+      }
+
+      // Slide up and fade in
+      tl.to(skylineRef.current, {
+        y: 10,
+        opacity: 1,
+        duration: 0.6,
+        ease: "power3.out",
+      }, "-=0.8");
+
+      // Pre-title and subtitle fade ins (and slide ups)
+      tl.fromTo(preTitleRef.current,
+        { y: 20, opacity: 0 },
+        { y: 0, opacity: 1, duration: 0.8, ease: "power3.out" },
+      "-=0.6");
+      tl.fromTo(subTitleRef.current,
+        { y: 20, opacity: 0 },
+        { y: 0, opacity: 1, duration: 0.8, ease: "power3.out" },
+      "-=0.6");
+
+      // Continuous floating effect starting after a delay
+      tl.to(logoRef.current, {
+        y: -10,
+        duration: 2,
+        ease: "power2.inOut",
+        yoyo: true,
+        repeat: -1
+      }, "+=0.5");
+    }, comp); // scope the selector to this component
+
+    return () => ctx.revert(); // Cleanup on unmount
   }, []);
 
+  const handleLogoClick = () => {
+    const newCount = clickCount + 1;
+    setClickCount(newCount);
+
+    if (newCount === 5) {
+      const ctx = gsap.context(() => {
+        gsap.killTweensOf(logoRef.current); // stop current floating animation
+        // teleportation animation timeline
+        const tl = gsap.timeline({
+            onComplete: () => {
+                setClickCount(0); 
+                // Resume the infinite floating
+                gsap.to(logoRef.current, {
+                    y: -10, duration: 2, ease: "power2.inOut", yoyo: true, repeat: -1
+                });
+            }
+        });
+        // Fly Down and Fade Out
+        tl.to(logoRef.current, {
+            x: -100,
+            y: 100,
+            opacity: 0,
+            duration: 0.5,
+            ease: "power2.in"
+        });
+        // Teleport to Top Right
+        tl.set(logoRef.current, {
+            x: window.innerWidth / 2,
+            y: -window.innerWidth / 2,
+        });
+        // Fly back to center
+        tl.to(logoRef.current, {
+            x: 0,
+            y: 0,
+            opacity: 1,
+            duration: 1.5,
+            ease: "power4.out"
+        });
+
+      }, comp);
+    }
+  };
+
   return (
-    <section ref={heroRef} className="relative h-screen bg-gradient-to-b from-[#0f0029] via-[#1f0140] to-[#fd9453] text-white overflow-hidden">
-      <StarryBackground />
-
-      <Navbar brandName="HackUTD" links={[]} />
-
-      <div className="absolute top-4 right-4 md:top-8 md:right-8 z-20 flex items-center gap-2 md:gap-4">
-        <a 
-          href="https://www.instagram.com/hackutd" 
-          target="_blank" 
-          rel="noopener noreferrer"
-          className="text-white hover:text-pink-400 transition-colors duration-300"
-        >
-          <svg className="w-5 h-5 md:w-6 md:h-6" fill="currentColor" viewBox="0 0 24 24">
-            <path d="M12 2.163c3.204 0 3.584.012 4.85.07 3.252.148 4.771 1.691 4.919 4.919.058 1.265.069 1.645.069 4.849 0 3.205-.012 3.584-.069 4.849-.149 3.225-1.664 4.771-4.919 4.919-1.266.058-1.644.07-4.85.07-3.204 0-3.584-.012-4.849-.07-3.26-.149-4.771-1.699-4.919-4.92-.058-1.265-.07-1.644-.07-4.849 0-3.204.013-3.583.07-4.849.149-3.227 1.664-4.771 4.919-4.919 1.266-.057 1.645-.069 4.849-.069zm0-2.163c-3.259 0-3.667.014-4.947.072-4.358.2-6.78 2.618-6.98 6.98-.059 1.281-.073 1.689-.073 4.948 0 3.259.014 3.668.072 4.948.2 4.358 2.618 6.78 6.98 6.98 1.281.058 1.689.072 4.948.072 3.259 0 3.668-.014 4.948-.072 4.354-.2 6.782-2.618 6.979-6.98.059-1.28.073-1.689.073-4.948 0-3.259-.014-3.667-.072-4.947-.196-4.354-2.617-6.78-6.979-6.98-1.281-.059-1.69-.073-4.949-.073zm0 5.838c-3.403 0-6.162 2.759-6.162 6.162s2.759 6.163 6.162 6.163 6.162-2.759 6.162-6.163c0-3.403-2.759-6.162-6.162-6.162zm0 10.162c-2.209 0-4-1.79-4-4 0-2.209 1.791-4 4-4s4 1.791 4 4c0 2.21-1.791 4-4 4zm6.406-11.845c-.796 0-1.441.645-1.441 1.44s.645 1.44 1.441 1.44c.795 0 1.439-.645 1.439-1.44s-.644-1.44-1.439-1.44z"/>
-          </svg>
-        </a>
-        <a 
-          href="https://www.linkedin.com/company/hackutd" 
-          target="_blank" 
-          rel="noopener noreferrer"
-          className="text-white hover:text-blue-400 transition-colors duration-300"
-        >
-          <svg className="w-5 h-5 md:w-6 md:h-6" fill="currentColor" viewBox="0 0 24 24">
-            <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/>
-          </svg>
-        </a>
+    <section
+      ref={comp}
+      className="relative min-h-screen bg-gradient-to-b from-[#050014] via-[#04002A] to-[#050032] text-white overflow-x-hidden overflow-y-visible"
+      >
+      <div className="absolute bottom-0 left-1/2 transform -translate-x-1/2 w-[120vw] h-[50vh] z-0">
+        <div
+          className="absolute inset-0 blur-3xl scale-150 opacity-70"
+          style={{
+            background:
+              "radial-gradient(ellipse 50% 110% at center bottom, rgba(255, 177, 94, 1) 40%, rgba(242, 88, 204, 0.9) 70%, rgba(231, 105, 199, 0.4) 85%, transparent 100%)",
+            pointerEvents: "none",
+          }}
+        />
       </div>
+      <Stars speed={1000} />
+      <div className="absolute top-1/3 left-0 right-0 transform -translate-y-1/2 z-10">
+        <div className="flex flex-col md:flex-row items-center justify-center space-x-4 px-4">
+          
+          <div onClick={handleLogoClick} className="cursor-pointer z-50">
+            <Image 
+                ref={logoRef}
+                src="/logo.svg" 
+                alt="HackUTD Logo" 
+                width={120} 
+                height={120} 
+                className="w-16 h-16 md:w-24 md:h-24 lg:w-32 lg:h-32 mb-2 md:mb-0 opacity-0 select-none" // added select-none to prevent highlighting
+                draggable={false}
+            />
+          </div>
 
-      <div className="z-10 relative flex flex-col md:flex-row items-center justify-center h-full px-4 md:px-8">
-        <div ref={logoRef} className="p-2 mb-4 md:mb-0">
-          <Image src="HackUTD Fire Logo.svg" alt="logo" width={130} height={130} className="w-20 h-20 md:w-32 md:h-32" />
-        </div>
-        <div className="z-10 relative flex flex-col items-center md:items-start justify-center text-center md:text-left">
-
-          <div ref={titleRef} className="mt-2">
-            <img src="/Title.svg" alt="HackUTD" className="w-auto h-16 sm:h-20 md:h-24 lg:h-32" />
+          <div ref={titleRef} className="font-['CeraPro'] ml-6 md:ml-0 opacity-0"> {/* initial opacity 0 for animation */}
+            <p ref={preTitleRef} className="text-base md:text-md font-medium text-white md:ml-1">We are</p>
+            <h1 className="text-6xl sm:text-7xl md:text-8xl lg:text-[6rem] font-bold leading-none">
+              <span className="text-white">Hack</span>
+              <span className="bg-gradient-to-r from-pink-500 to-orange-400 text-transparent bg-clip-text">UTD</span>
+            </h1>
+            <p ref={subTitleRef} className="text-sm sm:text-base md:text-sm font-medium text-white ml-1 md:ml-0 mr-1 text-left md:text-right">
+              North America&apos;s Largest 24-hour Hackathon
+            </p>
           </div>
         </div>
       </div>
-
-        <div 
-          ref={skylineRef}
-          className="absolute bottom-0 left-0 right-0 z-50 opacity-0 translate-y-8"
-        >
-          <img src="/Skyline.svg" alt="Skyline" className="w-full h-auto" />
+      <div className="absolute left-0 right-0 z-10" style={{ lineHeight: 0, fontSize: 0, bottom: 0, marginBottom: 0 }}>
+        <div className="relative w-full" style={{ lineHeight: 0, fontSize: 0, marginBottom: 0 }}>
+          <div 
+            ref={skylineRef}
+            className="w-full scale-[2.2] md:scale-100 origin-bottom opacity-0 translate-y-10" // initial opacity 0 for animation
+            style={{ 
+              lineHeight: 0, 
+              margin: 0, 
+              padding: 0,
+              display: 'block',
+              transformOrigin: 'bottom center',
+              fontSize: 0,
+              marginBottom: 0
+            }}
+          >
+            <img
+              src="/dallas_skyline.svg"
+              alt="Dallas Skyline"
+              className="w-full h-auto object-cover block"
+              style={{ 
+                display: 'block', 
+                margin: 0, 
+                padding: 0, 
+                lineHeight: 0, 
+                verticalAlign: 'bottom', 
+                fontSize: 0,
+                marginBottom: 0
+              }}
+            />
+          </div>
         </div>
+        <div className="bg-black h-20 w-full -mt-12 md:-mt-16 relative z-10"></div>
+      </div>
     </section>
   );
 }
